SOL-150796: consolidate operator-facing banners into internal/banner package
Operator-facing banners were spread across two packages: the Hop 1
startup banner in internal/auth/banner.go, and the V1
OAuth-not-supported banner (just landed in commit 3 of this PR) inside
internal/config/config.go. The Hop 1 / Hop 2 alignment banner that
ships in the next commit would have made it three homes.

Consolidates everything into a new internal/banner package. Emitters
take primitives (strings, ints) rather than *config.ServerConfig, so
the package has no internal dependencies and neither config nor auth
hits an import cycle calling into it.

Mechanical relocation, no behavior change:

  - internal/banner/banner.go (new): consts + exported emitters
    LogStartupAuthMode(mode, issuer string)
    LogOAuthNotSupported(n int)
  - internal/banner/banner_test.go (moved from internal/auth/): tests
    updated to call the new primitive-arg signatures
  - internal/auth/banner.go: deleted
  - internal/auth/banner_test.go: deleted
  - internal/config/config.go: oauthNotSupportedBanner const +
    logOAuthNotSupportedBanner emitter removed; validate() now calls
    banner.LogOAuthNotSupported(oauthBrokerCount)
  - cmd/server/main.go: auth.LogStartupBanner(cfg) replaced with
    banner.LogStartupAuthMode(cfg.MCPClientAuth.Mode,
    cfg.MCPClientAuth.Issuer); banner import added
  - internal/auth/middleware.go: stale doc-comment reference updated

End-to-end verified: ran the server against a disabled-mode config and
against an oauth-broker config; the WARN startup banner and the
ERROR V1 not-yet-supported banner appear byte-identical to before
the refactor.

Rationale recorded in
docs/superpowers/plans/oauth-token-exchange/SOL-150796-T2-config-schema.md.
