+++ internal/banner/banner_test.go
@@ -12,15 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package auth
+package banner
 
 import (
 	"bytes"
 	"log/slog"
 	"strings"
 	"testing"
-
-	"github.com/SolaceDev/solace-broker-mcp/internal/config"
 )
 
 // captureSlog replaces slog.Default with a TextHandler writing to buf.
@@ -34,11 +32,10 @@ func captureSlog(t *testing.T) (*bytes.Buffer, func()) {
 	return buf, func() { slog.SetDefault(prev) }
 }
 
-func Test_StartupBanner_Disabled(t *testing.T) {
+func Test_StartupAuthMode_Disabled(t *testing.T) {
 	buf, restore := captureSlog(t)
 	defer restore()
-	cfg := &config.ServerConfig{MCPClientAuth: config.MCPClientAuthConfig{Mode: config.AuthModeDisabled}}
-	LogStartupBanner(cfg)
+	LogStartupAuthMode("disabled", "")
 	out := buf.String()
 	for _, want := range []string{
 		"level=WARN",
@@ -54,11 +51,10 @@ func Test_StartupBanner_Disabled(t *testing.T) {
 	}
 }
 
-func Test_StartupBanner_Static(t *testing.T) {
+func Test_StartupAuthMode_Static(t *testing.T) {
 	buf, restore := captureSlog(t)
 	defer restore()
-	cfg := &config.ServerConfig{MCPClientAuth: config.MCPClientAuthConfig{Mode: config.AuthModeStatic, DevToken: "x"}}
-	LogStartupBanner(cfg)
+	LogStartupAuthMode("static", "")
 	out := buf.String()
 	for _, want := range []string{
 		"level=WARN",
@@ -74,14 +70,10 @@ func Test_StartupBanner_Static(t *testing.T) {
 	}
 }
 
-func Test_StartupBanner_OAuth(t *testing.T) {
+func Test_StartupAuthMode_OAuth(t *testing.T) {
 	buf, restore := captureSlog(t)
 	defer restore()
-	cfg := &config.ServerConfig{MCPClientAuth: config.MCPClientAuthConfig{
-		Mode:   config.AuthModeOAuth,
-		Issuer: "https://idp.example.com",
-	}}
-	LogStartupBanner(cfg)
+	LogStartupAuthMode("oauth", "https://idp.example.com")
 	out := buf.String()
 	if !strings.Contains(out, "level=INFO") {
 		t.Errorf("expected INFO log for oauth mode, got: %s", out)
